@@ -1,14 +1,12 @@
-'use strict';
+import process from 'node:process';
+import {EventEmitter} from 'node:events';
+import itchy from 'itchy';
+import currify from 'currify';
+import spawnify from 'spawnify';
 
-const process = require('node:process');
-const {EventEmitter} = require('node:events');
-const itchy = require('itchy');
-const currify = require('currify');
-
-const spawnify = require('spawnify');
 const isString = (a) => typeof a === 'string';
 
-module.exports = (cmd, directories) => {
+export default (cmd, directories) => {
     const emitter = new EventEmitter();
     
     check(cmd, directories);
